Added function type checking. So now the program verifies the correct # of parameters are inputted for a function as well as their types.
Also added block scope, so things like if, while, else statements get their own local table entry now for when they declare variables.

What's left:
change type verification for function calls and parameter counting. Currently gcb(a-b,b) is seen as 4 parameters as I could items not commas (oops). Also a-b is seen as a type mismatch with gcd and - as a-b is not seen as one statement but multiple. So still needs tunning but very close.